requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Templating
jinja2==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
testcontainers==4.9.0
testcontainers[postgres]==4.9.0

# Utilities
python-dateutil==2.9.0
